Split primitives entities into separate files
